restructure protoParser and added nextKey support
