+++ proto/rainsMsg.capnp.go
@@ -1981,10 +1981,11 @@ const (
 	Obj_value_Which_regt    Obj_value_Which = 9
 	Obj_value_Which_infra   Obj_value_Which = 10
 	Obj_value_Which_extra   Obj_value_Which = 11
+	Obj_value_Which_next    Obj_value_Which = 12
 )
 
 func (w Obj_value_Which) String() string {
-	const s = "nameip6ip4redirdelegnamesetcertserviceregrregtinfraextra"
+	const s = "nameip6ip4redirdelegnamesetcertserviceregrregtinfraextranext"
 	switch w {
 	case Obj_value_Which_name:
 		return s[0:4]
@@ -2010,6 +2011,8 @@ func (w Obj_value_Which) String() string {
 		return s[46:51]
 	case Obj_value_Which_extra:
 		return s[51:56]
+	case Obj_value_Which_next:
+		return s[56:60]
 
 	}
 	return "Obj_value_Which(" + strconv.FormatUint(uint64(w), 10) + ")"
@@ -2369,6 +2372,36 @@ func (s Obj_value) NewExtra() (PublicKey, error) {
 	return ss, err
 }
 
+func (s Obj_value) Next() (PublicKey, error) {
+	p, err := s.Struct.Ptr(0)
+	return PublicKey{Struct: p.Struct()}, err
+}
+
+func (s Obj_value) HasNext() bool {
+	if s.Struct.Uint16(4) != 12 {
+		return false
+	}
+	p, err := s.Struct.Ptr(0)
+	return p.IsValid() || err != nil
+}
+
+func (s Obj_value) SetNext(v PublicKey) error {
+	s.Struct.SetUint16(4, 12)
+	return s.Struct.SetPtr(0, v.Struct.ToPtr())
+}
+
+// NewNext sets the next field to a newly
+// allocated PublicKey struct, preferring placement in s's segment.
+func (s Obj_value) NewNext() (PublicKey, error) {
+	s.Struct.SetUint16(4, 12)
+	ss, err := NewPublicKey(s.Struct.Segment())
+	if err != nil {
+		return PublicKey{}, err
+	}
+	err = s.Struct.SetPtr(0, ss.Struct.ToPtr())
+	return ss, err
+}
+
 // Obj_List is a list of Obj.
 type Obj_List struct{ capnp.List }
 
@@ -2420,132 +2453,138 @@ func (p Obj_value_Promise) Extra() PublicKey_Promise {
 	return PublicKey_Promise{Pipeline: p.Pipeline.GetPipeline(0)}
 }
 
-const schema_fb2d77234707241e = "x\xda\xacX\x7f\x8cTW\x15>\xe7\xde\x99}\xc3\xee" +
-	",;\xb3o6Q\xffp\x0d\xa5\x89\xdd\x84\x16wk" +
-	"\x936&tGW\xcbR\xca^\x06\"lh\xc2\xdb" +
-	"\x99\xcb\xec[f\xdf\x0c\xef\xbd\xa5;\x0d\x04\x9a\xd0H" +
-	"\x0d\xc6j\xbb\x86FH\xbaI\xb7\x96\xa6DP\xaa\x12" +
-	"1\xa2)\xb165\x94\xa6UQPI!\xf8\x83F" +
-	"l\x8a\xe2\xa2>s\xde\xbc_3;\x0bK\xf4\xafy" +
-	"s\xcfw\xdf;\xe7;\xe7}\xe7\xdc\xb7\xbc\xd0\xf2 " +
-	"\xfbT\xfcr\x0b\x80\x18\x8a\xb78\x0f}0t\xe2d" +
-	"v`\x0f\x88.D\xe7\xe3K\x95/\xdc\xf1\xd8\xb2\x1b" +
-	"\x10g\x0a@\xdfF>\x88\xea8W\x00T\x9d\x7f\x1b" +
-	"\xd0\xe98vp\xe0\xbb\x17g\xf7\x11\x9aG\xd0H\x10" +
-	"\x8c\xcd\xaa\xed1\xda\xb7(\xd6\x8d\x80\xce\xc7v\xaf\xe9" +
-	"\xfa\xe9\x07S/\x13\x9aE\xd0.\xe6\xce\xf8 \xaa\xf7" +
-	"\xc7\xe9\xf2\xd3\xf1/\x12\xfc\xc573wm\xfe\xc1{" +
-	"\xdf\x83t\x176\xa2\xa7Z\xb2\xa8\xce\xb4\xd0c\xa6[" +
-	".\x03:\xcb\xbaN\xef\xfd\xcd\xe3]?j\x06>\xa8" +
-	"t\xa2zX!\xf0!\x85\xc0\x7fM\xfc\xf2\x99\xfe+" +
-	"\xbf:\xd9\x00\xa6\xdb\xf5\xbd\x9a\xe8D\xf5T\x82.\x7f" +
-	"\x92p\xfd\xf8\xfd\x84\x18\xcf\xee\xbb\xef\xcd\xa6n\xdf\xdf" +
-	"\xda\x89\xea\xcaV\xba\x1chu\xe1'>\xf9\xa5%\xf2" +
-	"\xb4v\xba)'\xd3m\xb3\xea\xe16B\x1fjs9" +
-	"\xf9\xea\xf1k\xab\xfa\xefx\xeft3\xbf\x7f\x9elE" +
-	"\xf5l\x92\xf6\xbd\x9b$\xbf\xff\x16\xdb\xf0v\xf2\xfbW" +
-	"\x7f\xdb\x0c|\xa6}\x0c\xd5K\xed\x04\xbe\xd0N\xe0\xd6" +
-	"\x1f\xae{\xf4\x99\xef\xfc\xee\x12\x88\x8f\":/\x17\x9e" +
-	"Z:\xf9\xf6=\x7f\x81\xf5\xa8`\x12\x99z~\xf1," +
-	"`\xdf\x85\xc5\xdfb\x80\xce7u\xe5\x897\xba\x8e\xfc" +
-	"\xb91\xeb\xe4s\xdfL\xba\x15\xd5W\xd3t\xe7\xa3i" +
-	"\xcazp\xaf&huu\xe7\x1b\xea\xc6N\xbaZ\xdf" +
-	"\xf9\x18\xa03\xf8\xe3\x9e\x95\xab\x0e\x1c\xbf\xd2\xe0\xb3\x9b" +
-	"\xe8\xa9\xce^Tg\x08\xdc7\xdd\xe9\xd2q\xee\xfc\xd8" +
-	"\x8b\xca7\xae~\xd8p\xeb\x01T\x12\x00}\xe7\xd5%" +
-	"\xa8\xbe\xaf\x12\xfeO\xea\xcf\x08\x7f\xa7\xb5\xd7\xbcW|" +
-	"\xfdFcn\\\xc7Ou=\x80\xea\xbb]\xe4\xcb\x99" +
-	"\xae\xcb\xb0\xcc\xa9\x98e\xbb|\x8f\xa9\xc5t\xc3Zm" +
-	"\x15\xef\xcek\x15\xa3\xf2\xc0#e[\xdf\xa2\xe75[" +
-	"/\x1b9\x99\xa7\x1f\x80!D\x91\xe41\x80\x18\x02\xa4" +
-	"\x07z\x01\xc4\x83\x1c\xc5\xc3\x0c\x113Hk+{\x00" +
-	"\xc4\xe78\x8a!\x86i\x86\x19d\x00\xe9\xd5\xb4\xf8\x10" +
-	"G\xb1\x8ea\xb7]\xde*\x0dl\x07\x86\xed\x80\x1dv" +
-	"\xb5\"1\x06\x0cc\x80\x1d\x05\xcd\xd60\x09\x0c\x93\x80" +
-	"\x81[\xac\xce\xad\xa1\x89\x91\x15%=\xbfJV\xc9\x97" +
-	"L\xe0\xcb\xceA\x00\xb1\x83\xa3\xd8\xcb0\xed;\xf3$" +
-	"=w7G\xb1\x8f!\xb2\x9a/O-\x01\x10{8" +
-	"\x8a\xa7\x19\xa69f\x90\x03\xa4\xbf2\x0c \xf6q\x14" +
-	"\xfb\x19\xa6c,\x831\x80\xf4\x14->\xcbQ<\xcf" +
-	"\xd0\xd9*\xab\xb9\x8a\x96\x97\x00\x10\xb8\x1b\xf5]\xd9*" +
-	"\xab~P\xcev\xad\xa4\x17r\xba\x01</1\x0e\x0c" +
-	"\xe3\xfe\xe2z\xc3\x06\xae\x97\x82\xc5\xe6\xdc\xf7\x17\x0a\xa6" +
-	"\xb4,1!\xcdj\x1d\xf7\x1f\x09\xe2}n\x04@\xec" +
-	"\xe7(^\x88\xc4;M\x099\xc0Q\xbcD\xe4{\x01" +
-	"\xcfd\x01\xc4\xf3\x1c\xc5+\x0c\x91\xd7\xe2=D\xc0\x17" +
-	"8\x8a#\x14/\xd6\xe2=L\xc0\x978\x8ac\x0c\xd3" +
-	"q\x9e\xc18@\xfa(-\xbe\xc2Q\xbc\xce\xd0\xb1&" +
-	"F\xc6d\xde\xee\x07\xa5P0\xfdD\xd5'tW\xbe" +
-	"l\xd8r\xd2\x0e\xad\xd5\x8a\xb4|\x96v\xc9\xc9\x8an" +
-	"J\xcb\x8f\x7fW\xb9B\xc1Y\xb8\x18p\x88\xa3\x0b[" +
-	"|+Z\x86\xcb\x86\xf4Y!RR\x01)\x1a\x91\xb2" +
-	"\x99\xa3(EH\xd1)\x8d\xa3\x1c\xc5\x9e\x08)Od" +
-	"#\xe5\xc2=V\x9e\xccz\xe5r`\xbeX\x1dK/" +
-	"\x1a\x9a=a\x02\x97\x81\xd3\xa9P\xdc\x00i\xb1\x91\x83" +
-	"\xda\x7f\xc3\x0e7\x04\x92U\xdb\x10\xc4\xcb\xeb\xe2]\xab" +
-	"\xd1_iYZ\x11eC\xa4\x94\xc0M\x1c\xc5h$" +
-	"R\x99\xf5\xc2\xdf\x11\x89\xb4J\xe1Or\x14\xcfF\"" +
-	"\xfd\xda\x18\x80x\xbaV\x13\xcd\xf2\x17\xf55\x10\x1f\xcf" +
-	"\xd7\x05Q\xe0\xe4\xb5\x8a6\xa2\x97t\xe8\xb0\xf5\x10\x97" +
-	"l\xc8o}\xbc\xb9Q\xcd,\xd42\x8bFC\xb9g" +
-	"\xbdw\xf1H$\xde\xc3\xc3am\x06\xf1\x9e\xa2\x1ax" +
-	"\x8d\xa3x+\x12\xef/h\xfb\xeb\x1c\xc5;T\xef\xb1" +
-	"Z\xbd\x9fY\x0b \xde\xe2(\xceQ\xbd\xc7k\xf5~" +
-	"\x96\x90\xefp\x14\x7f`M\xb8\x08d\xfb\xb6\xb8\xf0J" +
-	"i\x18\x94\xb2!\xeb\x8b\",\x12\xc7\xd4\x8c\xa2\xfc\xbc" +
-	"Y\x06\x1c\x0f0\xee\xda\xba\xf2\x1cM\xac\xe7-\xa2\x13" +
-	"sx\xeb\x0d5,\xe0\xed`OD;|\xde\xa6\xb3" +
-	"\xa1v\xf821\xd3\x13JG \x13\x87\xb2\x11\xed\xf0" +
-	"e\"\xd0\x8e\xd7\xe6(\xbc\xa1\x8d\xcf\x1bt\x9d\x84\xde" +
-	"\xae8\xd47\x86\x9c^\\\xe1\xe6B.\xa41\xac\x8d" +
-	"4\x01\xbfK\xd55\x01\xcej\x14D\x9b\x00z\x85\x13" +
-	"\xe5\xafYcp\xb4R\xb1l\xea\xf6(\xa5\xd2_[" +
-	"pO\xa8uA\xbf\x954\xcfxT\x02\x1b5p8" +
-	"\"\x02~\xbc\xd5\x11O\x04nS\x03\xff\x7f\xe5}k" +
-	"]\xb9\xa9\xe6\xf7[\x964\xbdQ\xa4\x83~\x17\"\xfc" +
-	"\x14J\x81\xa3\xd8\x1d\x09z\xe7\xb0\x17\xf4\xfeH\xd0S" +
-	"YO\x0e\xe7\x15\xfe\xb9\xee\x07\xe3\xde\xedH\xc1\x9c\x97" +
-	"\xbey-\xaf\x19\x19[q\xf7v\xad4!\xc5r\x1e" +
-	"K:\x8e\x1b\x90\xda\x8f=\x00\xb9\xcf \xc7\xdc\x06d" +
-	"\xd8\x8e\xffq\xdc\xa0\xd4\xf5\xb8\x04 7D\x86Md" +
-	"`\xffv\xdc\xc0\xd4\x8d\xaea\x1d\x196\x93\x81\xff\xcb" +
-	"qKX}\x14{\x01r\x1b\xc8P C\xec\x86\xe3" +
-	"\xbe\xc8\xaa\xe6\x1a6\x91a\x94\x0c\xf1Y'\x83-\x00" +
-	"\xaa\xc4,@n3\x19Jdh\xf9\xa7\x93qG[" +
-	"\xdd\xf5\xaa@\x86\x0a\x19\x94\xebN\x06\x13\x00\xea\xb8\xbb" +
-	"c\x94\x0c6\x19\x12\xffp2\xb8\x08@\xdd\xe6\xee(" +
-	"\x91a\x92\x0c\x8b\xfe\xeed\xb0\x15@\x9dp\x0d\x152" +
-	"\xec C\xeb5'\x83m\x00j\xd5\xf5\xca&\xc3n" +
-	"2\xb4}\xe8d0\x09\xa0\xeet\x0d\x93d\xd8\x83\xcc" +
-	"\x13\x9b\xfaf\xa3\xe8\x95\xfb|\xbe\x15\xbdro0\x99" +
-	"\x98\xb2\xa0\x87SLA\x96d\x11S\xe1\xe1\x0d\x10S" +
-	"\x80\xbb\xe8\x8e\x96\x0c\x15+/M\x1bS\xe1\x88\xed\xc1" +
-	",in\xd7\xf3\x12S\xe1\xa9\xa1f\xe90e1x" +
-	"\x0a\xfd\x09G#\xdd\xd8bjs\x1f\xd9-'\xedf" +
-	"\xeb\xf3u\xcd\xda\x93W\x1a[\xcas\x06\xf4\x9e\x85\x0e" +
-	"\xe8$\x92\x0fs\x14\x1bX\xbd`wT\xca\xa6\x8d\x0a" +
-	"0T\\\x07t\xd2\xb5*i]\x02\x18&nV\xc1" +
-	"5g\x12\x813w\xd1\x83\x97r\x14\xcb\xc9\x19\x8c\x1c" +
-	"\xc5\xd2\xcbz\x81\xd5\xb5\x82n\xb7\xf8\xe7\x098\xa2\x05" +
-	"\xddy_\x0b\"\x82?\xd2L\xf0#\xb2\x16h\xc1s" +
-	"\xc3\x9e\x8e\x1f\x8bh\xc1Q\xda~\x84\xa38\x11\x19\x15" +
-	"\x8e\xd3\xf6c\x1c\xc5\xc9P \x1e\x01eNg\xfb\x9f" +
-	"\x05b\x81\xb3Bsbj\xa3\xa2\xdb\x18\x14\x8f\x96O" +
-	"\xb8\xf2\xe1\xf2\x12\x1dwj\xd2A\xabg{\xc3y\xa7" +
-	"\xa6\x1b\xc4\xccy\xca\xd5\xaf9\x8a\x8b\xbeh\x105\x17" +
-	"\x08{\x8e\xa3\xf8\xa3\xaf\x18\xc4\xcd%\x9a%/r\x14" +
-	"W}\xb9\xa0\x81\xe0}Z\xbd\xc2Q\\\xf7\xb5\xa2\x05" +
-	" }\xed\xcb\x00\xe2:\xc7\\,P\x0a\xf7\x9b\x08\x8e" +
-	"\x00\xac\xa5\x978\x89\x0c\x1d\xcdK0\xa0\xd10s\xd1" +
-	"\xdba\xd1\x94\x88\xa9\xf0s\x85\xf7\x9e=N\x84\xa5\xc2" +
-	"O\x07\x1e|\x1b\x0dG\x98\x0a\xbfWx/\x93\xe1\x9d" +
-	"o\xc1m'\x98\x0a\xbf\xf1x\x00\xcd;\x84A\x87w" +
-	"\x87\xe0CM=\xa0\x1f\xfdz\x04h\x18\xed#(/" +
-	"\xa1\xa9\xf0\xfb\xcdM_\xea\xcf\xd2\x1d\xdd\xe3\xb7\\\xb1" +
-	"\xc6\xad\x88\x86\x86\xd7\xd3l\xfe\xef\x8dtA\xbf\xc8\xf5" +
-	"A\xef\xf8c\x87s\xdd6\xda]\xe2(&Y\xfd\x08" +
-	"\xd6=A\x05\x14\xcc,\xa3\x9a5\xda_*\x96\xa3\x87" +
-	"\xde\xe8t\xf2\xdf\x00\x00\x00\xff\xffK\x18u\x18"
+func (p Obj_value_Promise) Next() PublicKey_Promise {
+	return PublicKey_Promise{Pipeline: p.Pipeline.GetPipeline(0)}
+}
+
+const schema_fb2d77234707241e = "x\xda\xacXml\x14\xd7\xd5>g\xee\xaeg\xb1w" +
+	"\xf1\xaeg\x8c\xde\xbc?\xea\x8a\x10\xa9X\"Im\x1a" +
+	")\xf9C\xbc\xad\xdb`\x92\xe0\xcb\x12\x95Z\x8d\x94\xf1" +
+	"\xee\xc5\x1e\xb3\xcc.3c\xe2\x8d@\x18\xc9\xa8\xa6r" +
+	"T\xd2\xe0\xca\xa8 \x81\x04Q\x88\x82JZ\xfa\x81J" +
+	"UZ%j\x1a\xa5\"D\xe9\x07-4\xb5\x02\xa2j" +
+	"\x89J\xd2\xd0R\xd3v\xaa3;_\xbb^\x13\xa3\xf6" +
+	"\x97\xd7\xf7\x9c;s\x9e\xe7\x9c\xfb\x9cs\xe7~\xab\xe9" +
+	"a\xe9\xd3\xf1\xabM\x00\xbc?\xde\xe4<\xf2a\xff\x99" +
+	"\xb3\xd9\xde\x09\xe0\xed\x88\xce'V\xc8_\xb8\xfb\xe9U" +
+	"\xb7 .\xc9\x00\xdd_b}\xa8le2\x80\xa2\xb3" +
+	"o\x01:\xad\xa7\x0e\xf5~\xe7\xf2\xdc\x14y\xb3\x887" +
+	"\x92\x0b\xc6\xe6\x94T\x8c\xf6-\x89u \xa0\xf3\xff\xe3" +
+	"\xeb\xdb\x7f\xfa\xe1\xf4K\xe4-E\xbc]\x9f{\xe2}" +
+	"\xa8<\x18\xa7\x9f\x9f\x89\x7f\x91\xdc_xS]\xf9\xd4" +
+	"\x0f\xde\xfb\x1ed\xda\xb1\xde{\xba)\x8b\xca\xb1&z" +
+	"\xcd\x91\xa6\xab\x80\xce\xaa\xf6s\x93\xbf}\xa6\xfdG\x8d" +
+	"\x9c\x0f\xc9m\xa8\x9c\x90\xc9\xf9\xb8L\xce\x7fI\xfc\xea" +
+	"\xf9\x9ek\xbf>[\xe7L\x8f\xeb\xfen\xa2\x0d\x95\xd7" +
+	"\x12\xf4\xf3'\x097\x8ewG\xf9\xd6\xec\xd4\x03o6" +
+	"\x0c\xfb\xc1\xe66T\xd66\xd3\xcf\xdef\xd7\xfd\xcc\xa7" +
+	"\xbe\xb2\\\x9c\xd3\xce5\xe4\xe4H\xcb\x9cr\xa2\x85\xbc" +
+	"\x8f\xb7\xb8\x9c|\xed\xf4\x8du=w\xbfw\xaeQ\xdc" +
+	"?O6\xa3r!I\xfb~\x99\xa4\xb8?\x88mz" +
+	";\xf9\xfd\xeb\xbfk\xe4|>5\x82\xca\x95\x149\xcf" +
+	"\xa6\xc8\xb9\xf9\x87\x1b\x9f|\xfe\xdb\xbf\xbf\x02\xfc.D" +
+	"\xe7\xa5\xc2\xde\x15co\xdf\xf7gx\x02eL\xa1\xa4" +
+	"\\Z:\x07\xd8=\xbb\xf4]\x09\xd0\xf9\xa6.\xef~" +
+	"\xa3\xfd\xe4\x9f\xea\xb3N1w\xcff\x9aQ\xf9 C" +
+	"O~?CY\x0f\x9e\xd5\xc0[\xd9\xdb\xf6\x862\xdd" +
+	"F\xbf\x9ek{\x1a\xd0\xe9\xfbq\xe7\xdau\x07O_" +
+	"\xab\x8b\xd9M\xf4\xf9\xb6.Tf\xc9\xb9\xfbR\x9bK" +
+	"\xc7\xc5K#/\xc8\xdf\xb8\xfeQ\xdd\xa3{QN\x00" +
+	"t\xdf\xa5.Ge\xa5\xea\x96\x8b\xfa3\xf2\xbf\xc7\x9a" +
+	"4W\xf3\xaf\xdf\xaa\xcf\x8d\x1b8.{\x08\x95\xcc2" +
+	"\x8a%\xb5\xec*\xacr\xcaf\xc9.\xddgj1\xdd" +
+	"\xb0\x1e\xb3\x86\xee\xcdke\xa3\xfc\xd0\xe3%[\xdf\xac" +
+	"\xe75[/\x199\x91\xa7?\x00\xfd\x88<\xc9b\x00" +
+	"1\x04\xc8\xf4v\x01\xf0\x87\x19\xf2G%DT\x91\xd6" +
+	"\xd6v\x02\xf0\xcf1\xe4\xfd\x12f$TQ\x02\xc8<" +
+	"F\x8b\x8f0\xe4\x1b%\xec\xb0K[\x84\x81)\x900" +
+	"\x05\xd8jW\xca\x02c a\x0c\xb0\xb5\xa0\xd9\x1a&" +
+	"A\xc2$`\x10\x96T\x13V\xff\xe8\xe0\x9a\xa2\x9e_" +
+	"'*\x14\x8b\x1a\xc4\xb2\xb3\x0f\x80\xef`\xc8'%\xcc" +
+	"\xf8\xc1\xec\xa1\xf7\x8e3\xe4S\x12\xa2T\x8de\xefr" +
+	"\x00>\xc1\x90\xef\x930\xc3PE\x06\x90yv\x00\x80" +
+	"O1\xe43\x12fb\x92\x8a1\x80\xcc4-\xeeg" +
+	"\xc8\x0fK\xe8l\x11\x95\\Y\xcb\x0b\x00\x08\xc2\x8d\xc6" +
+	".o\x11\x15\x1f\x94\xb3]+\xea\x85\x9cn\x00\xcb\x0b" +
+	"\x8c\x83\x84q\x7f\xf1\x09\xc3\x06\xa6\x17\x83\xc5\xc6\xdc\xf7" +
+	"\x14\x0a\xa6\xb0,>*\xccJ\x0d\xf7\xff\x17\xe0=0" +
+	"\x08\xc0g\x18\xf2\xa3\x11\xbcG(!\x07\x19\xf2\x17\x89" +
+	"|\x0f\xf0\xb1,\x00?\xcc\x90\xbf,!\xb2*\xde\xe3" +
+	"\xe4x\x94!?Ix\xb1\x8a\xf7\x049\xbe\xc8\x90\x9f" +
+	"\x920\x13g*\xc6\x012\xaf\xd0\xe2\xcb\x0c\xf9\xeb\x12" +
+	":\xd6\xe8\xe0\x88\xc8\xdb= \x17\x0a\xa6\x9f\xa8\xda\x84" +
+	"\xee\xca\x97\x0c[\x8c\xd9\xa1\xb5R\x16\x96\xcf\xd2.1" +
+	"V\xd6Ma\xf9\xf8w\x95\xca\x04\xce\xc2\xa5\x80\xfd\x0c" +
+	"]\xb7\xa5\x1fG\xcb@\xc9\x10>+DJ: E" +
+	"#R\x9eb\xc8\x8b\x11RtJ\xe30C>\x11!" +
+	"ew6R.\xccceO\xd6+\x97\x83\x0bau" +
+	",}\xc8\xd0\xecQ\x13\x98\x08\x82N\x87\xe2\x06H\x8b" +
+	"\xf5\x1cT\xff7\xecpC Y\xd5\x0d\x01^V\x83" +
+	"w\x83F\xff\x0a\xcb\xd2\x86P\xd4!\xa5\x04~\x99!" +
+	"\x1f\x8e \x15Y\x0f\xfe\x8e\x08\xd2\x0a\xc1\x1fc\xc8\xf7" +
+	"G\x90>7\x02\xc0\xf7Uk\xa2Q\xfe\xa2\xb1\x06\xe2" +
+	"\xe3\xc5\xba(\x0a\x9c\xbcV\xd6\x06\xf5\xa2\x0e\xad\xb6\x1e" +
+	"\xfa%\xeb\xf2[\x8b77\xac\x99\x85jf\xd1\xa8+" +
+	"\xf7\xacw\x16OF\xf0\x9e\x18\x08k3\xc0\xfb\x1a\xd5" +
+	"\xc0\xab\x0c\xf9[\x11\xbc\xbf\xa0\xed\xaf3\xe4\xefP\xbd" +
+	"\xc7\xaa\xf5~~\x03\x00\x7f\x8b!\xbfH\xf5\x1e\xaf\xd6" +
+	"\xfb\x05\xf2|\x87!\xff\x83\xd4\x80\x8b@\xb6\xef\x88\x0b" +
+	"\xaf\x94\x06@.\x19\xa2\xb6(\xc2\"qL\xcd\x18\x12" +
+	"\x9f7K\x80[\x03\x1fwmci\x9e&\xd6\xf2\x16" +
+	"\xd1\x89y\xbcu\x85\x1a\x16\xf0v\xa83\xa2\x1d>o" +
+	"G\xb2\xa1v\xf82q\xac3\x94\x8e@&\x8eg#" +
+	"\xda\xe1\xcbD\xa0\x1d\xaf\xceSxC\xdb\xba \xe8\x1a" +
+	"\x09\xbdSq\xa8m\x0c9}h\x8d\x9b\x0b\xb1\x98\xc6" +
+	"\xb0!\xd2\x04\xfc.U\xd3\x04\x98T\xa5 \xda\x04\xd0" +
+	"+\x9c(\x7f\x8d\x1a\x83\xa3\x15\x87J\xa6n\x0fS*" +
+	"\xfd\xb5E\xf7\x84j\x17\xf4[I\xe3\x8cG%\xb0^" +
+	"\x03\x07\"\"\xe0\xe3\xad\x0cz\"p\x87\x1a\xf8\xbf+" +
+	"\xef\x8f\xd7\x95\xdbj~\x8fe\x09\xd3\x1bEZ\xe9\xef" +
+	"b\x84\x9f\xa0\x14\x18\xf2\xf1\x08\xe8\x9d\x03\x1e\xe8\x99\x08" +
+	"\xe8\xe9\xac'\x87\x0b\x0a\xff\xfc\xf0\x83q\xefN\xa4`" +
+	"\xde\xa1o\\\xcb\xeb\x07G\xd6\xdc\xbb]+\x8e\x0a\xbe" +
+	"\x9a\xc5\x92\x8e\xe3\x02R\x9e\xc4N\x80\xdc&d\x98\xb3" +
+	"Q\xc2\x14\xfe\xdbqA)\xdbp9@\xaeH\x861" +
+	"2H\xffr\\`\xca\xa8k(\x93a\x07\x19\xd8?" +
+	"\x1d\xb7\x84\x95\x0av\x01\xe4l2\x8c\x93!v\xcbq" +
+	"\x0f\xb2\xb2\xd35\x8c\x91a\x82\x0c\xf19G\xc5&\x00" +
+	"e7f\x01r;\xc80I\x86\xa6\x7f8\xaa;\xda" +
+	"\xeeq\xa3\x1a'\xc3\x14\x19\xe4\x9b\x8e\x8a\x09\x9ay\xdd" +
+	"\x1d\x13d\xd8G\x86\xc4\xdf\x1d\x15\x97\x00(\xcf\xba;" +
+	"&\xc9\xb0\x9f\x0cK\xfe\xe6\xa8\xd8L\xb3\xb1k\x98\"" +
+	"\xc3\x0c\x19\x9ao8*\xb6\x00(\xd3nT\xfb\xc8p" +
+	"\x90\x0c-\x1f9*&\x01\x94\x03\xaea?\x19\x0e\x93" +
+	"!\xf9WG\xc5\x14\x80r\xc8}\xd4\x0c\x19\x8e\xa2\xe4" +
+	"\xa9Pm\x17\x92\xf5\xf2\x03~\"d\xbd\xbc:\x18Y" +
+	"LQ\xd0\xc3\xf1\xa6 \x8ab\x08\xd3\xe1\xad\x0e\x10\xd3" +
+	"\x80\xbb\xe8\x89\x96\x08\xa5,/L\x1b\xd3\xe1\xec\xed\xb9" +
+	"Y\xc2\xdc\xae\xe7\x05\xa6\xc3\xebD\xd5\xd2j\x8a\xa1\xe0" +
+	"-\xf4O83\xe9\xc6fS\x9b\xff\xca\x0e1f7" +
+	"Zo5\xa8\xa4\xe6-/\xd4e\xab\x01\xad56\x97" +
+	"\xe6\x0d\xf4\x9d\x8b\x1d\xe8IT\x1fe\xc87I\xb5\x02" +
+	"\xdfZ.\x996\xca \xa1\xec\x06\xa0\x93\x0eVH\x1b" +
+	"\x13 a\xe2v\x15_\x0d&\x11\x04\xb3\x92^\xbc\x82" +
+	"!\xbf\x9f\x82\xc1\xc8\xd5-\xb3\xaa\x0b\xa4\x9a\xd6\xd1\xe1" +
+	"\x1e\x96\x05\x00G\xb4\xa3#\xefkG\xa4A\x0c6j" +
+	"\x10\x11\x19\x0c\xb4\xe3\xc0\x80\xa7\xfb\xa7\"\xda\xf1\x0am" +
+	"?\xc9\x90\x9f\x89\x8c\x16\xa7i\xfb)\x86\xfcl((" +
+	"\x8f\x83<\xaf\x13\xfe\xd7\x82\xb2\xc8\xd9\xa211\xd5\xd1" +
+	"\xd2m$\xb2G\xcb']\xb9qy\x89\x8eGU\xa9" +
+	"\xa1\xd5\x0b]\xe1|T\xd5\x19b\xe6\x12\xe5\xea7\x0c" +
+	"\xf9e_d\x88\x9aY\xf2\xbd\xc8\x90\xff\xd1W\x18\xe2" +
+	"\xe6\x0a\xcd\x9e\x97\x19\xf2\xeb\xbe\xbc\xd0\x00\xf1>\xad^" +
+	"c\xc8o\xfa\xda\xd2\x04\x90\xb9\xf1U\x00~\x93a." +
+	"\x16(\x8b\xfb\x0d\x05\x07\x016\xd0\xd9N\xa2\x84\x8e\xe6" +
+	"%\x18\xd0\xa8\x9b\xd1\xe8\xd0X4Ub:\xfc\xbc\xe1" +
+	"\x1d\x9ag\x88\xb0t\xf8\xa9\xc1s\xdfF\xc3\x14\xa6\xc3" +
+	"\xef\x1b\xdea2\xbc\xfb0\xb8\xed\x07\xd3\xe17!\xcf" +
+	"A\xf3.m\xd0\xea=!\xf8\xb0S\xeb\xd0\x83~=" +
+	"\x02\xd4]\x05\"^^B\xd3\xe1\xf7\x9e\xdb\x1e\xea\xcf" +
+	"\xd2\x13\xdd\xeb\xbaX\xb3\xde\xad\x88\xba\x06\xd9\xd9\xe8\xbe" +
+	"\xd0\x15\xe9\x9a~\x91\xeb}\xdeu\xc9\x0e\xe7\xc0m\xb4" +
+	"\xbb\xc8\x90\x8fI\xb5#[\xc7(\x15P0\xe3\x0ck" +
+	"\xd6pOq\xa8\x14\xbd$G\xa7\x99\xff\x04\x00\x00\xff" +
+	"\xff%\xc2\x83\x84"
 
 func init() {
 	schemas.Register(schema_fb2d77234707241e,
